修改实体类,完善新建订单功能,订单结算,取消订单,确认收货功能

+++ src/main/java/com/qf/bean/Users.java
@@ -17,7 +17,7 @@ public class Users {
 
     private Date birthday;
 
-    private Double emoney;
+    private Integer emoney;
 
     private Integer userstatus;
 
@@ -81,11 +81,11 @@ public class Users {
         this.birthday = birthday;
     }
 
-    public Double getEmoney() {
+    public Integer getEmoney() {
         return emoney;
     }
 
-    public void setEmoney(Double emoney) {
+    public void setEmoney(Integer emoney) {
         this.emoney = emoney;
     }
